+++ lib/flutter_tex.dart
@@ -1,5 +1,7 @@
 library flutter_tex;
 
+export 'src/utils/enums.dart';
 export 'src/utils/tex_view_child.dart';
+export 'src/utils/tex_view_decoration.dart';
 export 'src/utils/tex_view_style.dart';
 export 'src/views/flutter_tex_view_controller.dart';