+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.5.0
+* API Changes.
+* `teXHTML` removed from API
+* `children` has been added, where you can pass List of `TeXViewChild`.
+* `TeXViewStyles` added where you can easily customise TeXView Widget.
+* `onTap` callback support for each individual child.
+
+
 ## 3.1.4+1
 * FlutterWeb rendering Bug fixed.